Fix numeric types in unit tests

+++ src/test/kotlin/radar/scene/Point2DTest.kt
@@ -10,19 +10,19 @@ class Point2DTest {
 
     @Test
     fun `test isInScene when point is within bounds`() {
-        val point = Point2D(x = 50f, y = 50f)
+        val point = Point2D(x = 50.0, y = 50.0)
         assertTrue(point.isInScene())
     }
 
     @Test
     fun `test isInScene when point is out of bounds negative`() {
-        val point = Point2D(x = -1f, y = -1f)
+        val point = Point2D(x = -1.0, y = -1.0)
         assertFalse(point.isInScene())
     }
 
     @Test
     fun `test isInScene when point is out of bounds positive`() {
-        val point = Point2D(x = GRID_SIZE_X.toFloat() + 1, y = GRID_SIZE_Y.toFloat() + 1)
+        val point = Point2D(x = GRID_SIZE_X + 1, y = GRID_SIZE_Y + 1)
         assertFalse(point.isInScene())
     }
 }
